Only run migration when start server
